Fix architecture filtering
The regex match ended up stringifying to "amd64,amd64" and "arm64,arm64"
when modified to match either of (amd64|arm64), changing the path of the
object as a result:
```
> String(images[0].Name.match(/arm64/))
'arm64'
> String(images[0].Name.match(/(arm64|amd64)/))
'arm64,arm64'
```
The image notifications now include an Architecture field, so use this
instead of matching against the image name.  The Architecture names
from Amazon do not match the names used in the FreeBSD images, so use
an object to lookup the arch name.  If the Architecture is in the map,
we have a match, otherwise the image is skipped.